Start separating generic and Theano-specific meta types and functionality
