+++ symbolic_pymc/tensorflow/__init__.py
@@ -1,0 +1,3 @@
+from tensorflow.python.framework import ops
+
+ops.disable_eager_execution()
