+++ symbolic_pymc/theano/ops.py
