+++ symbolic_pymc/unify.py
@@ -1,25 +1,17 @@
 import reprlib
 
-import theano.tensor as tt
-
 from functools import wraps
 
 from multipledispatch import dispatch
 
 from kanren import isvar
 from kanren.term import term, operator, arguments
-from kanren.facts import fact
-from kanren.assoccomm import commutative, associative
-
-# from kanren.goals import LCons, _lcons_unify
 
 from unification.more import unify
 from unification.core import reify, _unify, _reify, Var
 
-from .meta import MetaSymbol, MetaVariable, mt, metatize
+from .meta import MetaSymbol, MetaVariable
 from .utils import _check_eq
-
-tt_class_abstractions = tuple(c.base for c in MetaSymbol.__subclasses__())
 
 etuple_repr = reprlib.Repr()
 etuple_repr.maxstring = 100
@@ -171,16 +163,6 @@ def unify_MetaSymbol(u, v, s):
 
 
 _unify.add((MetaSymbol, MetaSymbol, dict), unify_MetaSymbol)
-_unify.add(
-    (MetaSymbol, tt_class_abstractions, dict), lambda u, v, s: unify_MetaSymbol(u, metatize(v), s)
-)
-_unify.add(
-    (tt_class_abstractions, MetaSymbol, dict), lambda u, v, s: unify_MetaSymbol(metatize(u), v, s)
-)
-_unify.add(
-    (tt_class_abstractions, tt_class_abstractions, dict),
-    lambda u, v, s: unify_MetaSymbol(metatize(u), metatize(v), s),
-)
 
 
 def _reify_MetaSymbol(o, s):
@@ -200,14 +182,6 @@ def _reify_MetaSymbol(o, s):
 
 
 _reify.add((MetaSymbol, dict), _reify_MetaSymbol)
-
-
-def _reify_TheanoClasses(o, s):
-    meta_obj = metatize(o)
-    return reify(meta_obj, s)
-
-
-_reify.add((tt_class_abstractions, dict), _reify_TheanoClasses)
 
 
 _isvar = isvar.dispatch(object)
@@ -254,7 +228,6 @@ def operator_MetaVariable(x):
 
 operator.add((MetaSymbol,), operator_MetaSymbol)
 operator.add((MetaVariable,), operator_MetaVariable)
-operator.add((tt.Variable,), lambda x: operator(metatize(x)))
 
 
 def arguments_MetaSymbol(x):
@@ -289,7 +262,6 @@ def arguments_MetaVariable(x):
 
 arguments.add((MetaSymbol,), arguments_MetaSymbol)
 arguments.add((MetaVariable,), arguments_MetaVariable)
-arguments.add((tt.Variable,), lambda x: arguments(metatize(x)))
 
 
 def _term_ExpressionTuple(rand, rators):
@@ -298,7 +270,6 @@ def _term_ExpressionTuple(rand, rators):
 
 
 term.add((object, ExpressionTuple), _term_ExpressionTuple)
-term.add((tt.Op, ExpressionTuple), lambda op, args: term(metatize(op), args))
 
 
 @dispatch(object)
@@ -337,11 +308,6 @@ def tuple_expression(x):
     return res
 
 
-@dispatch(tt_class_abstractions)
-def tuple_expression(x):
-    return tuple_expression(mt(x))
-
-
 @_reify.register(ExpressionTuple, dict)
 def _reify_ExpressionTuple(t, s):
     """When `kanren` reifies `etuple`s, we don't want them to turn into regular `tuple`s.
@@ -369,10 +335,5 @@ def _reify_ExpressionTuple(t, s):
     res = etuple(*res)
     return res
 
-
-fact(commutative, mt.add)
-fact(commutative, mt.mul)
-fact(associative, mt.add)
-fact(associative, mt.mul)
 
 __all__ = ["debug_unify", "etuple", "tuple_expression"]
