+++ tests/theano/test_linalg.py
@@ -6,9 +6,9 @@ import numpy as np
 
 from theano.gof.graph import inputs as tt_inputs
 
-from symbolic_pymc import NormalRV, observed
-from symbolic_pymc.meta import mt
-from symbolic_pymc.opt import eval_and_reify_meta, FunctionGraph
+from symbolic_pymc.theano.random_variables import NormalRV, observed
+from symbolic_pymc.theano.meta import mt
+from symbolic_pymc.theano.opt import eval_and_reify_meta, FunctionGraph
 from symbolic_pymc.unify import (etuple, tuple_expression)
 from symbolic_pymc.relations.linalg import (normal_normal_regression, buildo,
                                             normal_qr_transform)
